sub2/backend/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
joblib==0.14.0
numpy==1.17.3
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.1
sqlparse==0.3.0
pandas==0.25.3
django-allauth==0.42.0
django-rest-auth==0.9.5
django-rest-knox ==4.1.0
scikit-surprise==1.1.1
surprise==0.1